Test time-shortcut-picker in its own spec not in bookmark-test.js
